% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxPolygon.R
\name{bbox_from_points}
\alias{bbox_from_points}
\title{Create an sf bounding box from a list of longitude and latitudes}
\usage{
bbox_from_points(list_lng, list_lat, crs = 4326)
}
\arguments{
\item{crs}{}
}
\value{

}
\description{
Create an sf bounding box from a list of longitude and latitudes
}
\examples{

bbox_from_points(c(45.6,44.2), c(-71.1,-73.1))
}
