% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxPolygon.R
\name{bbox_from_vector}
\alias{bbox_from_vector}
\title{Create an sf bounding box from a vector with (xmin,ymin,ymax,ymax)}
\usage{
bbox_from_vector(
  v,
  crs,
  x_name_min = "xmin",
  y_name_min = "ymin",
  x_name_max = "xmax",
  y_name_max = "ymax"
)
}
\arguments{
\item{v}{vector with names (xmin,ymin,ymax,ymax)}

\item{x_name_min}{character}

\item{y_name_min}{character}

\item{x_name_max}{character}

\item{y_name_max}{character}
}
\value{
sf object representing bbox
}
\description{
This works automatically with raster extents or sf::bbox
}
\examples{
\dontrun{
library(sf)

shp_trees <-  st_read('https://www.donneesquebec.ca/recherche/dataset/bc5afddf-9439-4e96-84fb-f91847b722be/resource/bbdca0dd-82df-42f9-845b-32348debf8ab/download/vdq-arbrepotentielremarquable.geojson')
rasterKDECentroids <- st_kde(shp_trees , cellsize = 0.001, bandwith =c(.001, .001 ) )
bbox_from_vector(extent(rasterKDECentroids) , crs=4326)

}
}
