% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbox_polygon}
\alias{bbox_polygon}
\title{Create a st_polygon from a bounding box}
\usage{
bbox_polygon(x, buffer = 0)
}
\arguments{
\item{x}{sf or sfc object that we want to bound within some box}

\item{buffer}{dist to add around the bounding box to avoid the weird boundary effects}
}
\value{
st_polygon forming a box
}
\description{
Really useful function to define bounding box enveloppe
}
\details{
See https://stackoverflow.com/questions/45719790/create-voronoi-polygon-with-simple-feature-in-r
}
\examples{
  shpUnionedSchoolds \%>\% st_centroid() \%>\% bbox_polygon()
}
