% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxPolygon.R
\name{bbox_validation}
\alias{bbox_validation}
\title{Util to check if the bbox polygon produced is valid or not}
\usage{
bbox_validation(shp_bbox)
}
\arguments{
\item{shp_bbox}{}
}
\value{
logical/boolean T for ok F for warning
}
\description{
Util to check if the bbox polygon produced is valid or not
}
