% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_turn_off_s2}
\alias{check_turn_off_s2}
\title{Code to check if s2 should be used and turn off UNLESS `USE_S2_SPHERICAL` is set}
\usage{
check_turn_off_s2()
}
\description{
Code to check if s2 should be used and turn off UNLESS `USE_S2_SPHERICAL` is set
}
