% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{create_opq_query_add_features}
\alias{create_opq_query_add_features}
\title{Add feature, value pairs recursively to an osm overpass query}
\usage{
create_opq_query_add_features(
  opq_query,
  list_osm_params = list(list(key = "highway"))
)
}
\arguments{
\item{list_osm_params}{}
}
\value{

}
\description{
Add feature, value pairs recursively to an osm overpass query
}
