% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createBuffer}
\alias{createBuffer}
\title{Create a buffer around an sf object}
\usage{
createBuffer(shp, dist)
}
\arguments{
\item{shp, }{sf object}

\item{dist, }{distance (supposed to be in some valid unit)}
}
\value{
shpBuffSchools
}
\description{
Create a buffer around an sf object
}
