% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{download_cd}
\alias{download_cd}
\title{Download census division from the stats can website & filter on province}
\usage{
download_cd(
  PR = "35",
  use_cartographic = F,
  dir_to_dl = here::here("data", "cd"),
  dir_cache = here::here("data", "cache"),
  new_crs = NULL
)
}
\arguments{
\item{PR}{str. province ID - e.g. 35 (On) or 24 (Qc)}

\item{use_cartographic}{bool}
}
\value{
shp_cd sf object with cd for that province - idname CDUID
}
\description{
Download census division from the stats can website & filter on province
}
\details{
Caches tmp files
}
