% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{download_econ_regions}
\alias{download_econ_regions}
\title{Download economic regions from the stats can website & filter on province}
\usage{
download_econ_regions(
  PR = "35",
  use_cartographic = F,
  dir_to_dl = here::here("data", "er"),
  dir_cache = here::here("data", "cache"),
  new_crs = NULL
)
}
\arguments{
\item{PR}{str. province ID - e.g. 35 (On) or 24 (Qc)}

\item{use_cartographic}{bool}
}
\value{
shp_econ_regions sf object with economic regions for that province
}
\description{
Download economic regions from the stats can website & filter on province
}
\details{
Caches tmp files
}
