% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{download_generic_wrappee}
\alias{download_generic_wrappee}
\title{Code logic for `download_econ_regions``}
\usage{
download_generic_wrappee(
  url_getter,
  PR = "35",
  use_cartographic = F,
  dir_to_dl = here::here("data", "cd"),
  new_crs = NULL
)
}
\description{
Code logic for `download_econ_regions``
}
