% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{download_prov}
\alias{download_prov}
\title{Download the province of quebec cartographic shp file + cache results}
\usage{
download_prov(
  use_cartographic = T,
  PR = "35",
  dir_to_dl = here::here("data", "pr"),
  dir_cache = here::here("data"),
  new_crs = NULL
)
}
\arguments{
\item{data_dir}{}
}
\value{
url (character)
}
\description{
Download the province of quebec cartographic shp file + cache results
}
