% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{get_cd_url}
\alias{get_cd_url}
\title{Get the stats can url for census division (2021)}
\usage{
get_cd_url(use_cartographic)
}
\arguments{
\item{use_cartographic}{}
}
\value{
url (character)
}
\description{
Get the stats can url for census division (2021)
}
