% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{get_econ_reg_url}
\alias{get_econ_reg_url}
\title{Get the stats can url for economic regions  (2021)}
\usage{
get_econ_reg_url(use_cartographic)
}
\arguments{
\item{use_cartographic}{}
}
\value{
url (character)
}
\description{
Get the stats can url for economic regions  (2021)
}
