% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{get_fsa_url}
\alias{get_fsa_url}
\title{Get FSA (first 3 digits of postal code) - 2021 data}
\usage{
get_fsa_url(use_cartographic)
}
\arguments{
\item{use_cartographic}{}
}
\description{
Get FSA (first 3 digits of postal code) - 2021 data
}
