% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{get_polygon_heatmap}
\alias{get_polygon_heatmap}
\title{Create polygons that represent a 2D density out of points or centroid of polygons}
\usage{
get_polygon_heatmap(df, crs_epsg = 4326, bw = 0.001, gsize = 500)
}
\arguments{
\item{df}{data.frame ish or sf}

\item{crs_epsg}{crs of the newly create polygon, default 4326 if inputing a df and not explicitely set}

\item{bw}{bandwidth  (same for both dimesions)}

\item{gsize}{gid size (same for both dimesions)}
}
\description{
If using a sf object, will use st_centroids
Otherwise, need to ensure lng and lat columns are available
}
