% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_polygons.R
\name{get_prov_url}
\alias{get_prov_url}
\title{Get the stats can url for 13 provinces & territories (2016)}
\usage{
get_prov_url(use_cartographic)
}
\arguments{
\item{use_cartographic}{}
}
\value{
url (character)
}
\description{
Uses the 2016 census dataset, but this is immaterial since no changes over time
}
