% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadShp.R
\name{get_zipped_remote_shapefile}
\alias{get_zipped_remote_shapefile}
\title{Download a (zipped) shp file from a url}
\usage{
get_zipped_remote_shapefile(url, dirToDownload = here::here("data"))
}
\arguments{
\item{dirToDownload}{}
}
\value{
shp downloaded shp file
}
\description{
Download a (zipped) shp file from a url
}
