% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids.R
\name{getCentroids}
\alias{getCentroids}
\title{Get the long/lat of centroids of sf object (adds 2 new columns if they do not already exist)}
\usage{
getCentroids(shp)
}
\arguments{
\item{shp, }{sf object}
}
\value{
shp with 2 new lng and lat columns
}
\description{
Similar to cbind( shp , st_coordinates(st_centroid(shp)))
}
