% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readShpFiles.R
\name{getShpLayerNameInDir}
\alias{getShpLayerNameInDir}
\title{Calls auxiliary GeneralHelpers functions to find a shp file in dir}
\usage{
getShpLayerNameInDir(abspath)
}
\arguments{
\item{abspath}{}
}
\description{
Calls auxiliary GeneralHelpers functions to find a shp file in dir
}
