% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{getVoronoiInterpolation}
\alias{getVoronoiInterpolation}
\title{Get a piecewise constant inerpolation based on the voronoi diagram}
\usage{
getVoronoiInterpolation(shp, varToInterpolate, fctToAggregate = mean)
}
\arguments{
\item{shp, }{sf object}

\item{varToInterpolate, }{chr variable to interpolate}

\item{fctToAggregate}{fct used e.g. mean, median, max or other statistic}
}
\value{
voronoi, sf object with col varToInterpolate
}
\description{
Get a piecewise constant inerpolation based on the voronoi diagram
}
