% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{is_within_cvx_hull}
\alias{is_within_cvx_hull}
\title{Check to see if a dataframe of points lies within the convex hull of another dataframe of points}
\usage{
is_within_cvx_hull(coords_to_check, coords_area)
}
\arguments{
\item{coords_area}{}
}
\value{

}
\description{
Geographic 4326 coordinates with X and Y columns for lon and lat for both dataframes
}
