% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{make_isochrone_new_point}
\alias{make_isochrone_new_point}
\title{Wrapper over make_isochrone that creates an isochrone based on the weighted graph computed and a starting point}
\usage{
make_isochrone_new_point(graph, df_start_point, limit_minutes = c(5, 10, 15))
}
\arguments{
\item{df_start_point}{}
}
\value{

}
\description{
Wrapper over make_isochrone that creates an isochrone based on the weighted graph computed and a starting point
}
