% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSfValidProjRmZ}
\alias{makeSfValidProjRmZ}
\title{Wrapper for makeValid, st_zm, st_transform ...}
\usage{
makeSfValidProjRmZ(shp)
}
\arguments{
\item{shp}{}
}
\description{
Wrapper for makeValid, st_zm, st_transform ...
}
