% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValid.R
\name{makeSfValidRmZM}
\alias{makeSfValidRmZM}
\title{Wrapper for makeValid, st_zm,}
\usage{
makeSfValidRmZM(shp)
}
\arguments{
\item{shp}{valid and cleaned sf object}
}
\value{
shp, cleaned shp
}
\description{
Wrapper for makeValid, st_zm,
}
