% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{osm_buffer_query_helper}
\alias{osm_buffer_query_helper}
\title{Get osm objects within a fixed radius of a given point}
\usage{
osm_buffer_query_helper(
  shp_point,
  buffer_km = 5,
  list_osm_params = list(list(key = "highway"))
)
}
\arguments{
\item{list_osm_params}{}
}
\value{
osm_results_streets osm results: list with linestrings, points and polygons + aux data
}
\description{
This is useful namely to get the street network around a point
}
