% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQcCity.R
\name{readQcCityData}
\alias{readQcCityData}
\title{Read all the shp files in file.path("Data", "Rdata", "GeoData","QuebecCityOpenData")}
\usage{
readQcCityData(readSP = F,
  qcDirPath = "/home/charles/Projects/SfSpHelpers")
}
\value{
a named list of sf objects
}
\description{
Read all the shp files in file.path("Data", "Rdata", "GeoData","QuebecCityOpenData")
}
\details{
names include:
"vdq-quartier",
"vdq-batiments",
"vdq-arbrepotentielremarquable",
"vdq-zonesaccessansfil"
}
