% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readShpFiles.R
\name{readShpFile}
\alias{readShpFile}
\title{Read the shp file with rgdal::readOGR}
\usage{
readShpFile(pathName, layerName, readSP)
}
\value{
shp, spatial polygons/point object
}
\description{
Data taken from https://www.donneesquebec.ca/recherche/fr/dataset/vque_9
}
