% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring_split.R
\name{remove_line_endpoints}
\alias{remove_line_endpoints}
\title{Removes the endpoints for a collection of LINESTRINGS}
\usage{
remove_line_endpoints(shp_lines, ...)
}
\arguments{
\item{shp_lines}{sf object with multiple rows}

\item{...}{parameters passed to to remove_line_endpoints_lwgeom}
}
\value{

}
\description{
Removes the endpoints for a collection of LINESTRINGS
}
