% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalePolygons.R
\name{scale_geom_sfc}
\alias{scale_geom_sfc}
\title{Scale an individual sfc}
\usage{
scale_geom_sfc(sf_to_scale, scale_factor = 0.75)
}
\arguments{
\item{scale_factor}{}
}
\value{

}
\description{
Scale an individual sfc
}
