% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{sf2graph}
\alias{sf2graph}
\title{Creates a weighted graph whose weight is given by the time required to travel along an edge given a certain mode}
\usage{
sf2graph(sfdf)
}
\arguments{
\item{sfdf}{}
}
\value{

}
\description{
Creates a weighted graph whose weight is given by the time required to travel along an edge given a certain mode
}
