% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfSpHelpers_package.R
\docType{package}
\name{sfSpHelpers}
\alias{sfSpHelpers}
\title{SfSpHelpers is an R package for routine spatial operations that can be reused in different contexts}
\author{
Charles Gauvin (@charles.gauvin123@gmail.com)
}
