% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring_split.R
\name{split_lines_remove_line_endpoints}
\alias{split_lines_remove_line_endpoints}
\title{Splits line into segments using SfSpHelpers::split_lines and then removes the first and last segments (if possible)}
\usage{
split_lines_remove_line_endpoints(
  shp_line,
  id_col,
  min_distance_m_to_remove = 10
)
}
\arguments{
\item{min_distance_m_to_remove}{}
}
\description{
Splits line into segments using SfSpHelpers::split_lines and then removes the first and last segments (if possible)
}
