% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFilters.R
\name{st_generic_filter}
\alias{st_generic_filter}
\title{Generic fct to filter a shp using a spatial predicate (e.g. st_intersect) and another shp}
\usage{
st_generic_filter(shpToFilter, shpToCompare, st_predicate)
}
\arguments{
\item{shpToFilter}{initial sf object to filter}

\item{shpToCompare}{sf object use for filtering}

\item{st_predicate}{spatial predicate - eg sf::st_within}
}
\value{
shpFiltered
}
\description{
Generic fct to filter a shp using a spatial predicate (e.g. st_intersect) and another shp
}
