% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{st_get_linestring_osm_by_mode}
\alias{st_get_linestring_osm_by_mode}
\title{Creates a graph with weights given by the mode chosen}
\usage{
st_get_linestring_osm_by_mode(
  osm_results_streets,
  osm_modes = c("motorcar", "foot", "bicycle")
)
}
\arguments{
\item{osm_modes}{}
}
\value{
osm list of sf LINESTRING
}
\description{
Creates a graph with weights given by the mode chosen
}
