% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFilters.R
\name{st_intersects_filter}
\alias{st_intersects_filter}
\title{Convenience wrapper that calls st_generic_filter with st_intersects}
\usage{
st_intersects_filter(shpToFilter, shpToCompare)
}
\arguments{
\item{st_predicate}{}
}
\description{
Convenience wrapper that calls st_generic_filter with st_intersects
}
