% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring.R
\name{st_linestring_add_endpoints}
\alias{st_linestring_add_endpoints}
\title{Add start and end points to a sf liestring object}
\usage{
st_linestring_add_endpoints(shp_segments)
}
\arguments{
\item{shp_segments}{, sf object}
}
\value{
shp_segments , sf object with 4 new columns start_x,start_y, end_x,end_y
}
\description{
Add start and end points to a sf liestring object
}
