% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring_endpoints.R
\name{st_linestring_to_sf_linestring_endpoints}
\alias{st_linestring_to_sf_linestring_endpoints}
\title{Uses st_linestring_add_endpoints to take a potentiall complicated linestring and convert it into a simple linestring based only on the endpoints of the initial linestring}
\usage{
st_linestring_to_sf_linestring_endpoints(shp_segments)
}
\arguments{
\item{shp_segments}{}
}
\value{
shp_segments_endpoints
}
\description{
Uses st_linestring_add_endpoints to take a potentiall complicated linestring and convert it into a simple linestring based only on the endpoints of the initial linestring
}
