% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{st_make_isochrone_concave_hull}
\alias{st_make_isochrone_concave_hull}
\title{Compute isochrones based on a cppRouting graph (whose edges have weight given by time in minutes)}
\usage{
st_make_isochrone_concave_hull(graph, from = "1", limit_minutes = c(5, 10, 15))
}
\arguments{
\item{lim}{}
}
\value{

}
\description{
The limits depends on the units used for the 'dist' column - for this function requires that the dist be in seconds
}
