% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{st_make_isochrone_main}
\alias{st_make_isochrone_main}
\title{Man function to call to get}
\usage{
st_make_isochrone_main(
  shp_point,
  buffer_km = 5,
  limit_minutes = c(5, 10, 15),
  list_modes = c("motorcar", "foot", "bicycle"),
  list_osm_params = list(list(key = "highway"))
)
}
\arguments{
\item{list_modes}{}
}
\value{
shp_iso_all_modes: sf POLYGON
}
\description{
Man function to call to get
}
