% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{st_make_isochrone_new_point}
\alias{st_make_isochrone_new_point}
\title{Wrapper over st_make_isochrone_concave_hull that creates an isochrone based on the weighted graph computed and a starting point}
\usage{
st_make_isochrone_new_point(
  graph,
  df_start_point,
  limit_minutes = c(5, 10, 15),
  walk_speed_kmh = 5
)
}
\arguments{
\item{df_start_point}{}
}
\value{

}
\description{
Wrapper over st_make_isochrone_concave_hull that creates an isochrone based on the weighted graph computed and a starting point
}
