% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmIsochrone.R
\name{st_osm_city_centroid_from_string}
\alias{st_osm_city_centroid_from_string}
\title{Simple helper to get the city centroid from a string (in 4326 geographic coordiantes)}
\usage{
st_osm_city_centroid_from_string(city_str)
}
\arguments{
\item{city_str}{}
}
\value{

}
\description{
Simple helper to get the city centroid from a string (in 4326 geographic coordiantes)
}
\examples{
st_osm_city_centroid_from_string('Atlanta')
}
