% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadShp.R
\name{st_read_remote}
\alias{st_read_remote}
\title{Fct that wraps st_read with a curl call}
\usage{
st_read_remote(url, dirToDownload = here("Data"))
}
\arguments{
\item{url}{url adress of shp to download}

\item{dirToDownload}{directory where the file will be unzipped , can be "", in which case a temp dir is used}
}
\value{
shp
}
\description{
Can read geojsons for instance
}
\details{
Tries to download the file to dirToDownload/nameOfFile for persistance.
If this fails, it just downloads it to a tmp file.
If dirToDownload/nameOfFile already exists, then it just reads back the shp file.
nameOfFile is created from the url path
}
