% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring_endpoints.R
\name{st_sf_linestring_from_points}
\alias{st_sf_linestring_from_points}
\title{From a df where each row represents an edge in 2D of the form (lng_1,lat_1),  (lng_2,lat_2), create an sf object with LINESTRING geometry}
\usage{
st_sf_linestring_from_points(
  df_segments,
  col_start_x = "start_x",
  col_start_y = "start_y",
  col_end_x = "end_x",
  col_end_y = "end_y",
  crs = 4326
)
}
\arguments{
\item{col_start_x}{start represents the first endpoint (start is a misnomer there is no edge orientation)}

\item{col_end_y}{}
}
\value{
shp_segments_endpoints sf object
}
\description{
From a df where each row represents an edge in 2D of the form (lng_1,lat_1),  (lng_2,lat_2), create an sf object with LINESTRING geometry
}
