% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCentroids}
\alias{getCentroids}
\title{Get the long/lat of centroids of sf object}
\usage{
getCentroids(shp)
}
\arguments{
\item{shp, }{sf object}
}
\value{
shp with 2 new lng and lat columns
}
\description{
Get the long/lat of centroids of sf object
}
