% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeValid}
\alias{makeValid}
\title{Make sure polygons that intersect are treated  MULTIPOLYGONS}
\usage{
makeValid(shp)
}
\arguments{
\item{shp, }{sf object to clean}
}
\value{
shp, valid sf object
}
\description{
Useful to deal with stats can shp files (which has many self-overlapping polygons)
}
