% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring_split.R
\name{remove_line_endpoints_lwgeom}
\alias{remove_line_endpoints_lwgeom}
\title{Get a substring from a linestring}
\usage{
remove_line_endpoints_lwgeom(
  shp_line,
  id_col,
  min_distance_m_to_remove = NULL,
  min_proportion_remove = NULL
)
}
\arguments{
\item{min_proportion_remove}{}
}
\value{

}
\description{
Wrapper over lwgeom::st_linesubstring with possibility to either use meters or
}
