% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialKMeans.R
\name{spatialKMeans}
\alias{spatialKMeans}
\title{Smooth out a large number of points or polygons by taking a spatially weighted k means}
\usage{
spatialKMeans(shp, numCentroids = NULL, propToKeep = NULL,
  numClosestPoints = 10, var = "ass_diff", aggFct = median)
}
\arguments{
\item{numClosestPoints}{- analoguous to bandwith - larger number reduces the variance}

\item{var}{- string - variable to summarise}

\item{aggFct}{- function to use to summarise - e.g. median or mean}
}
\value{
shpCentroids with column var
}
\description{
Smooth out a large number of points or polygons by taking a spatially weighted k means
}
