% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{st_kde}
\alias{st_kde}
\title{Produce a 2D kernel density estimate based on some POINTS geometry}
\usage{
st_kde(points, cellsize, bandwith, extent = NULL)
}
\arguments{
\item{points}{sf object with POINT geometry}

\item{cellsize}{size of cell}

\item{bandwith}{bandwith for 2D kde, larger: smoother, smaller: spikier}

\item{extent}{allows cropping to a specific subregion of the map}
}
\value{
raster
}
\description{
Note: Taken from https://github.com/r-spatial/sf/issues/1201
}
\examples{
\dontrun{
librarry(sf)
libary(here)
library(magrittr)
library(dplyr)

shpBuildings <- SfSpHelpers::get_zipped_remote_shapefile ("https://data.montreal.ca/dataset/4ad6baea-4d2c-460f-a8bf-5d000db498f7/resource/866a3dbc-8b59-48ff-866d-f2f9d3bbee9d/download/uniteevaluationfonciere.geojson.zip")
shpCentroids <- shpBuildings \%>\% st_centroid

rasterKDECentroids <- st_kde(shpCentroids,0.01,0.01)
plot(rasterKDECentroids)
}
}
