% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring_endpoints.R
\name{st_linestring_to_sf_linestring_endpoints}
\alias{st_linestring_to_sf_linestring_endpoints}
\title{Convenience function over  st_linestring_add_endpoints and st_sf_linestring_from_points
Takes a potentially complicated linestring and converts it into a simple linestring based only on the *endpoints* of the initial linestring
(keeps the geometry as LINESTRING)}
\usage{
st_linestring_to_sf_linestring_endpoints(shp_segments)
}
\arguments{
\item{shp_segments}{}
}
\value{
shp_segments_endpoints
}
\description{
Convenience function over  st_linestring_add_endpoints and st_sf_linestring_from_points
Takes a potentially complicated linestring and converts it into a simple linestring based only on the *endpoints* of the initial linestring
(keeps the geometry as LINESTRING)
}
